Add tests for instantiating a game
This, at the moment, is basically just doing the same sort of
thing that we already had with the VariableSubstitution visitor.

A couple of benefits: now we have a transformer (so no mutability)
which is safer. The transformer allows me to swap out variables
for other Nodes, like the literal done in the test case. And it
allows me to control how the recursion might go down, which I'll
need for things like FieldAccess where all I really want to do is
tunnel to the left hand side of the tree (or, perhaps, I should
try to make it parse so that FieldAccesses are right-oriented),
which would make it easier to get at the base.

Also it took me forever to figure out the typing because I was
trying to model CRTP from C++, instead of just using the Self
parameter.