Get inlining for scheme method calls working
Testing out "GeneralDoubleOTUC.proof". Now, the original game will
unfold so that the two have roughly the same lines. But the problem: we
need reordering. To get the reduction to work, we need to delay the
generation of the second key. Which is fine from a code equivalency
standpoint, it's just the proof engine isn't smart enough to recognize
that. So that's the next goal. Coerce these two games to evaluate as the
same. Will try to do so by collapsing unnecessary variables, figuring
out some sort of "canonical ordering", etc.